requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
instagrapi==1.16.30
Pillow==9.3.0
pycryptodomex==3.15.0
pydantic==1.10.2
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
typing-extensions==4.4.0
urllib3==1.26.13
